Filter highlights based on visible annotations in the sidebar
When the 'filter_highlights' feature flag is enabled, add and remove
highlights in the document to match the currently visible set of
annotations in the sidebar.

This means that when a search query is used to filter the visible
annotations, that affects what is shown in the page _as well as_ what is
shown in the sidebar.

This implementation is the simplest possible approach - highlights for
hidden annotations are just removed from the document and re-anchored if
the annotations are later shown. After we've done UX and performance
testing, we can implement something smarter to avoid the cost of
anchoring each time an annotation is shown.